raspberry/requirements.txt
requests==2.31.0
pydantic==2.7.3
numpy==1.26.4
opencv-python-headless==4.8.1.78
Pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
facenet-pytorch==2.5.3
RPi.GPIO==0.7.1
pydantic-settings==2.1.0
onnxruntime==1.19.0
gpiod==1.5.4
